Added Timers, made the return type of TimeHelper.getTime() "long" and finished the Animation and KeyFrame classes.

+++ src/mEngine/util/time/TimeHelper.java
@@ -23,7 +23,7 @@ public class TimeHelper {
 
     }
 
-    public static float getTime() {
+    public static long getTime() {
         return System.nanoTime() / 1000000;
     }
